refactor: Move guild model to guild package
