+++ common/src/main/java/com/wynntils/commands/OnlineMembersCommand.java
@@ -11,9 +11,9 @@ import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.consumers.commands.Command;
-import com.wynntils.models.players.type.GuildInfo;
-import com.wynntils.models.players.type.GuildMemberInfo;
-import com.wynntils.models.players.type.GuildRank;
+import com.wynntils.models.guild.type.GuildInfo;
+import com.wynntils.models.guild.type.GuildMemberInfo;
+import com.wynntils.models.guild.type.GuildRank;
 import com.wynntils.utils.mc.McUtils;
 import java.util.Arrays;
 import java.util.Collections;
