+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -29,6 +29,7 @@ import com.wynntils.models.elements.ElementModel;
 import com.wynntils.models.emeralds.EmeraldModel;
 import com.wynntils.models.gear.GearModel;
 import com.wynntils.models.gear.SetModel;
+import com.wynntils.models.guild.GuildModel;
 import com.wynntils.models.horse.HorseModel;
 import com.wynntils.models.ingredients.IngredientModel;
 import com.wynntils.models.inventory.InventoryModel;
@@ -41,7 +42,6 @@ import com.wynntils.models.npc.NpcModel;
 import com.wynntils.models.npcdialogue.NpcDialogueModel;
 import com.wynntils.models.objectives.ObjectivesModel;
 import com.wynntils.models.players.FriendsModel;
-import com.wynntils.models.players.GuildModel;
 import com.wynntils.models.players.PartyModel;
 import com.wynntils.models.players.PlayerModel;
 import com.wynntils.models.profession.ProfessionModel;
