+++ common/src/main/java/com/wynntils/features/ExtendedSeasonLeaderboardFeature.java
@@ -12,9 +12,9 @@ import com.wynntils.core.text.StyledText;
 import com.wynntils.handlers.labels.event.LabelIdentifiedEvent;
 import com.wynntils.handlers.labels.event.TextDisplayChangedEvent;
 import com.wynntils.handlers.labels.type.LabelInfo;
-import com.wynntils.models.players.label.GuildSeasonLeaderboardLabelInfo;
-import com.wynntils.models.players.profile.GuildProfile;
-import com.wynntils.models.players.type.GuildLeaderboardInfo;
+import com.wynntils.models.guild.label.GuildSeasonLeaderboardLabelInfo;
+import com.wynntils.models.guild.profile.GuildProfile;
+import com.wynntils.models.guild.type.GuildLeaderboardInfo;
 import com.wynntils.utils.StringUtils;
 import com.wynntils.utils.colors.ColorChatFormatting;
 import java.util.HashMap;
