+++ common/src/main/java/com/wynntils/features/chat/GuildRankReplacementFeature.java
@@ -15,7 +15,7 @@ import com.wynntils.core.text.StyledText;
 import com.wynntils.core.text.StyledTextPart;
 import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
 import com.wynntils.handlers.chat.type.RecipientType;
-import com.wynntils.models.players.type.GuildRank;
+import com.wynntils.models.guild.type.GuildRank;
 import com.wynntils.utils.type.IterationDecision;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
