+++ common/src/main/java/com/wynntils/models/guild/type/GuildInfo.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.players.type;
+package com.wynntils.models.guild.type;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
