+++ common/src/main/java/com/wynntils/models/guild/type/GuildLeaderboardInfo.java
@@ -2,6 +2,6 @@
  * Copyright © Wynntils 2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.players.type;
+package com.wynntils.models.guild.type;
 
 public record GuildLeaderboardInfo(int position, String guildName, long rating) {}
