+++ common/src/main/java/com/wynntils/models/guild/type/GuildRank.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.players.type;
+package com.wynntils.models.guild.type;
 
 public enum GuildRank {
     RECRUIT("Recruit", ""),
