+++ common/src/main/java/com/wynntils/models/players/type/WynnPlayerInfo.java
@@ -10,6 +10,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.wynntils.core.components.Models;
+import com.wynntils.models.guild.type.GuildInfo;
+import com.wynntils.models.guild.type.GuildMemberInfo;
+import com.wynntils.models.guild.type.GuildRank;
 import com.wynntils.utils.mc.McUtils;
 import java.lang.reflect.Type;
 import java.time.Instant;
